Модуль 7 Задание 1
Выполнила задание
