+++ src/resizer.js
@@ -345,5 +345,5 @@
     this.y = y;
   };
 
-  window.Resizer = Resizer;
+  module.exports = Resizer;
 })();
